requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
python-docx>=1.0.0
reportlab>=4.0.0
google-generativeai>=0.3.0
language-tool-python>=2.7.0
python-dotenv>=1.0.0
